feat: move component AccessDenied.tsx file from /components/Projects to /components
